all component files converted to TS
