+++ src/pages/modules/MainPhase/submodules/DataForm/DataForm.tsx
@@ -1,6 +1,4 @@
-import React, {
-  useState, useEffect, useCallback, useContext,
-} from 'react';
+import React, { useState, useEffect, useCallback, useContext } from 'react';
 // import PropTypes from 'prop-types';
 import {
   MEDIA_STATUS_COLORS,
@@ -13,74 +11,116 @@ import {
 import generateQueryJson from '../../../util/generateQueryJson';
 
 import './DataForm.css';
-import GlobalContext from '../../../util/GlobalContext.tsx';
+import GlobalContext, { GlobalContextOptions } from '../../../util/GlobalContext';
 import presets from '../../../Alert/presets';
+import { KeyPress } from 'interfaces/interfaces';
+
+interface DFProps {
+  title?: string;
+  image?: string;
+  color?: string;
+  type?: string;
+  token?: string;
+  mediaId?: number;
+  transitionCallback: Function;
+  presetProgress?: number;
+  presetScore?: number;
+}
 
 const DataForm = ({
-  title, image, color, type, token, mediaId, transitionCallback, presetProgress = undefined, presetScore = undefined,
-}) => {
+  title,
+  image,
+  color,
+  type,
+  token,
+  mediaId,
+  transitionCallback,
+  presetProgress = undefined,
+  presetScore = undefined,
+}: DFProps): React.ReactElement => {
   const [status, setStatus] = useState('CURRENT');
   const [progress, setProgress] = useState(presetProgress);
   const [score, setScore] = useState(presetScore);
 
-  const { setGlobalValues } = useContext(GlobalContext);
+  const { setGlobalValues }: GlobalContextOptions = useContext(GlobalContext);
 
   /* Get current media status color or default to black */
   const currentMediaColor = useCallback((override = status) => MEDIA_STATUS_COLORS[override] || '#222', [status]);
 
   const handleBadRequest = useCallback(() => {
-    setGlobalValues({
-      type: 'ALERT',
-      data: {
-        active: true,
-        content: 'Something went wrong with the request. The page will automatically refresh in 2 seconds.',
-        duration: 2000,
-        style: presets.red,
-      },
-    });
+    setGlobalValues &&
+      setGlobalValues({
+        type: 'ALERT',
+        data: {
+          active: true,
+          content: 'Something went wrong with the request. The page will automatically refresh in 2 seconds.',
+          duration: 2000,
+          style: presets.red,
+        },
+      });
     setTimeout(() => {
       window.location.reload();
     }, 2000);
   }, [setGlobalValues]);
 
-  const switchMediaMode = useCallback((newStatus) => {
-    setStatus(newStatus);
-    setGlobalValues({
-      type: 'ALERT',
-      data: {
-        active: true,
-        content: MEDIA_STATUS_ALERT_MESSAGES[newStatus],
-        style: {
-          border: `1px solid ${currentMediaColor(newStatus)}`,
-          backgroundColor: `${currentMediaColor(newStatus)}3`,
-        },
-      },
-    });
-  }, [currentMediaColor, setGlobalValues]);
+  const switchMediaMode = useCallback(
+    newStatus => {
+      setStatus(newStatus);
+      setGlobalValues &&
+        setGlobalValues({
+          type: 'ALERT',
+          data: {
+            active: true,
+            content: MEDIA_STATUS_ALERT_MESSAGES[newStatus],
+            style: {
+              border: `1px solid ${currentMediaColor(newStatus)}`,
+              backgroundColor: `${currentMediaColor(newStatus)}3`,
+            },
+          },
+        });
+    },
+    [currentMediaColor, setGlobalValues],
+  );
 
   /* Key Press Event Handler */
-  const handleKeyPress = useCallback((e) => {
-    const { [e.key.toLowerCase()]: newStatus } = {
-      u: 'CURRENT', c: 'COMPLETED', d: 'DROPPED', h: 'PAUSED',
-    };
-    if (newStatus) {
-      switchMediaMode(newStatus);
-    } else if (e.key === 'Enter') {
-      const options = generateQueryJson(POST_MEDIA_CHANGE_QUERY_GEN({
-        mediaId, status, score, progress,
-      }), token);
-      fetch(ANILIST_BASE_URL, options)
-        .then((resp) => resp.json())
-        .then((resp) => (resp.ok
-          ? transitionCallback() // return to search phase
-          : handleBadRequest()))
-        .catch(handleBadRequest);
-    }
-  }, [handleBadRequest, mediaId, progress, score, status, switchMediaMode, token, transitionCallback]);
+  const handleKeyPress = useCallback(
+    (e: KeyPress) => {
+      const keyPressed: string = e.key;
+      const { [keyPressed.toLowerCase()]: newStatus }: any = {
+        u: 'CURRENT',
+        c: 'COMPLETED',
+        d: 'DROPPED',
+        h: 'PAUSED',
+      };
+      if (newStatus) {
+        switchMediaMode(newStatus);
+      } else if (e.key === 'Enter') {
+        const options = generateQueryJson(
+          POST_MEDIA_CHANGE_QUERY_GEN({
+            mediaId,
+            status,
+            score,
+            progress,
+          }),
+          token,
+        );
+        fetch(ANILIST_BASE_URL, options)
+          .then(resp => resp.json())
+          .then(resp =>
+            resp.ok
+              ? transitionCallback() // return to search phase
+              : handleBadRequest(),
+          )
+          .catch(handleBadRequest);
+      }
+    },
+    [handleBadRequest, mediaId, progress, score, status, switchMediaMode, token, transitionCallback],
+  );
 
   /* Event handling setup and initial form focus */
   useEffect(() => {
-    document.getElementById('data-form-media-count-value').focus();
+    const mediaCountElem = document.getElementById('data-form-media-count-value');
+    mediaCountElem && mediaCountElem.focus();
     document.addEventListener('keydown', handleKeyPress);
 
     return () => document.removeEventListener('keydown', handleKeyPress);
@@ -103,9 +143,7 @@ const DataForm = ({
           style={{
             // TODO
             // backgroundImage: `url('${image}')`,
-            backgroundColor: `#${`${Math.floor(
-              Math.random() * 9,
-            )}`.repeat(6)}`,
+            backgroundColor: `#${`${Math.floor(Math.random() * 9)}`.repeat(6)}`,
             border: `1px solid ${currentMediaColor()}`,
           }}
         />
@@ -121,7 +159,7 @@ const DataForm = ({
               id="data-form-media-count-value"
               type="number"
               value={progress}
-              onChange={(e) => setProgress(e.target.value)}
+              onChange={e => setProgress(+e.target.value)}
               placeholder={`${MEDIA_TYPE_SINGLETON_TERM[type]}s`}
               style={{
                 border: `1px solid ${currentMediaColor()}`,
@@ -133,7 +171,7 @@ const DataForm = ({
                 id="data-form-score-value"
                 type="number"
                 value={score}
-                onChange={(e) => setScore(e.target.value)}
+                onChange={e => setScore(+e.target.value)}
                 placeholder="Score"
                 style={{
                   border: `1px solid ${currentMediaColor()}`,
@@ -141,7 +179,6 @@ const DataForm = ({
                 }}
               />
             ) : null}
-
           </div>
         </div>
       </div>
