+++ src/pages/modules/MainPhase/submodules/HelpMessage/HelpMessage.tsx
@@ -11,16 +11,18 @@ const HelpMessage = ({ substate, prevSubstate, helpMap }) => {
     transform: `translateX(${substate === 'TRANSITION' ? -20 : 0}px)`,
   });
 
-  return (helpMap[substate] || helpMap[prevSubstate]) && (
-    <animated.div style={helpMessageProps} id="help-message" className={substate}>
-      <IoIosInformationCircleOutline
-        size="1.75em"
-        style={{
-          paddingRight: '3px',
-        }}
-      />
-      {helpMap[substate] || helpMap[prevSubstate]}
-    </animated.div>
+  return (
+    (helpMap[substate] || helpMap[prevSubstate]) && (
+      <animated.div style={helpMessageProps} id="help-message" className={substate}>
+        <IoIosInformationCircleOutline
+          size="1.75em"
+          style={{
+            paddingRight: '3px',
+          }}
+        />
+        {helpMap[substate] || helpMap[prevSubstate]}
+      </animated.div>
+    )
   );
 };
