+++ tsconfig.json
@@ -7,6 +7,7 @@
       "dom.iterable",
       "esnext"
     ],
+    "downlevelIteration": true,
     "allowJs": true,
     "skipLibCheck": true,
     "esModuleInterop": true,
@@ -19,6 +20,7 @@
     "isolatedModules": true,
     "noEmit": true,
     "jsx": "react",
+    // Fix before full release
     "noImplicitAny": false
   },
   "include": [
